langbot/requirements.txt
Flask==2.0.3
transformers==4.20.1
nltk==3.7
langdetect==1.1.0
google-cloud-translate==2.0.1
torch==1.12.1